Made changes to CInstrument to support effects
I made what should be a non-breaking change to CInstrument. Each
instrument needs to know what 'level' it needs (0.0 to 1.0) for each
effects channel, so I added an array and a getter and setter for these
values. The number of effects channels is set at compile time by a macro
for the metaprocessor I set in resource.h. Again, this shouldn't have
any effect on your guys' code, but let me know if I've broke something.